+++ README.md
@@ -90,9 +90,9 @@
     - [x] 주문 금액이 100,000원 미만인 경우 배송비는 3,000원으로 설정된다.
     - [x] 주문 금액이 100,000원 이상인 경우 배송비는 0원으로 설정된다.
 
-  - [] PageTest
-    - [] url이 "/"인 경우, "SHOP"이 보여진다.
-    - [] url이 "/order-check"인 경우, "결제하기"가 보여진다.
+  - [x] PageTest
+    - [x] / 경로로 접근하면 SHOP이 보인다.
+    - [x] /order-check 경로로 접근하면 결제하기가 보인다.
 
 - [x] Routing
